pyproject.toml
[project]
name = "dataminds-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.39.14",
    "fastapi>=0.116.1",
    "python-dotenv>=1.1.1",
    "strands-agents>=1.1.0",
    "strands-agents-builder>=0.1.7",
    "strands-agents-tools>=0.2.2",
    "streamlit>=1.47.1",
]
